pyproject.toml
[tool.poetry]
name = "import-candidates"
version = "0.1.0"
description = ""
authors = ["Neoklis Kollias <neokliskol@protonmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.32.3"
uvicorn = "^0.32.0"
fastapi = "^0.115.2"
python-multipart = "^0.0.12"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
